requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
boto==2.48.0
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
coverage==4.3.1
cryptography==2.2.2
dj-database-url==0.4.2
Django==2.0.5
django-appconf==1.0.1
django-common-helpers==0.9.1
django-compat==1.0.14
django-cron==0.5.0
django-dbbackup==3.0.2
django-nose==1.4.4
django-storages==1.6.5
django-summernote==0.8.6
easy-thumbnails==2.3
ez-setup==0.9
factory-boy==2.8.1
Faker==0.7.3
gunicorn==19.6.0
idna==2.6
lxml==4.0.0
nose==1.3.7
paramiko==2.4.1
pilkit==1.1.13
pyasn1==0.4.2
pycparser==2.18
PyMySQL==0.8.0
PyNaCl==1.2.1
python-dateutil==2.6.0
python-docx==0.8.6
python-gnupg==0.4.2
python-resize-image==1.1.11
pytz==2018.3
reportlab==3.4.0
requests==2.18.4
six==1.10.0
sorl-thumbnail==12.3
urllib3==1.22
whitenoise==3.2.1